rename "importer" as DatabaseBuilder, and import command as build-flat
