+++ ifncli/managers/export/db/builder/__init__.py
@@ -1,6 +1,6 @@
 from .schema import SurveySchema
 from .version_selector import VersionSelector, VersionSelectorParser
-from .importer import Importer
+from .builder import DatabaseBuilder
 from .base import Writer, PrintWriter, SourceDataLoader
-from .profile import ImporterProfile
+from .profile import BuilderProfile
 from . import fake
